refactor: restDocs 위치 수정
refactor: restDocs 위치 수정
