+++ src/main/java/com/team/comma/domain/playlist/recommend/controller/RecommendController.java
@@ -3,6 +3,7 @@ package com.team.comma.domain.playlist.recommend.controller;
 import com.team.comma.domain.playlist.recommend.dto.RecommendListRequest;
 import com.team.comma.domain.playlist.recommend.dto.RecommendRequest;
 import com.team.comma.domain.playlist.recommend.service.RecommendService;
+import com.team.comma.domain.track.track.service.TrackService;
 import com.team.comma.global.common.dto.MessageResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -16,6 +17,12 @@ import javax.security.auth.login.AccountException;
 public class RecommendController {
 
     private final RecommendService recommendService;
+    private final TrackService trackService;
+
+    @GetMapping("/tracks")
+    public MessageResponse findTrackByMostFavorite() {
+        return trackService.findTrackByMostFavorite();
+    }
 
     @PostMapping
     public ResponseEntity<MessageResponse> recommendPlaylist(
