+++ src/main/java/com/team/comma/domain/track/track/controller/TrackController.java
@@ -20,9 +20,4 @@ public class TrackController {
         return trackService.findTrackByFavoriteTrack(accessToken);
     }
 
-    @GetMapping("/favorites")
-    public MessageResponse findTrackByMostFavorite() {
-        return trackService.findTrackByMostFavorite();
-    }
-
 }
